my-requirements.txt
numpy==1.18.1
pandas==1.0.4
scikit-learn==0.23.1
scipy==1.4.1
spacy==2.1.0
dill==0.2.9
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz
jedi==0.17.2
nlg-eval @ git+https://github.com/Maluuba/nlg-eval.git@4855e70301dcfb1d4943eb6c4d1213201e4f4b98
nlp==0.3.0
nltk==3.4.5
rouge-score==0.0.4
oauthlib==3.1.0
packaging==20.4
pickleshare==0.7.5
progress==1.5
python-dateutil==2.8.1
regex==2018.7.11
seaborn==0.10.1
segtok==1.5.10
Send2Trash==1.5.0
sentence-transformers==0.3.2
six==1.15.0
smart-open==2.0.0
sentencepiece==0.1.91
tensorboard==2.2.2
ujson==4.0.0
tqdm==4.46.1
tokenizers==0.8.1rc1
toolz==0.10.0
torch==1.5.1
torchtext==0.3.1
tensorboard-plugin-wit==1.6.0.post3
sqlitedict==1.6.0
srsly==1.0.2
statsmodels==0.11.1
tabulate==0.8.7
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1